Add     Rulet Rotate

+++ src/Engine2D/EngineObject.java
@@ -4,9 +4,9 @@ import UnityMath.Vector2;
 
 public abstract class EngineObject {
     public int id;//id
-    public int angX;//angle of X axis
-    public int angY;//angle of Y axis
-    public int angZ;//angle of Z axis
+    public double angX;//angle of X axis
+    public double angY;//angle of Y axis
+    public double angZ;//angle of Z axis
     public Vector2 center;//center
     public Vector2 position;//position
 }
